renamed applicationReducer back to userReducer to minimize changes
